nc_modernist/requirements.txt
attrs==18.1.0
click==6.7
google-api-python-client==1.7.4
